+++ mlfinance/nlp/bert.py
@@ -2,349 +2,61 @@
 import pandas as pd
 import numpy as np
 
-from tqdm.auto import tqdm
-
-import torch
-import torch.nn as nn
-from torch.utils.data import Dataset, DataLoader
-
-from transformers import (
-    BertTokenizerFast as BertTokenizer,
-    BertModel,
-    AdamW,
-    get_linear_schedule_with_warmup,
-)
-
 import pytorch_lightning as pl
-from pytorch_lightning.metrics.functional import accuracy, f1, auroc
-from pytorch_lightning.callbacks import ModelCheckpoint, EarlyStopping
-from pytorch_lightning.loggers import TensorBoardLogger
 
-from sklearn.model_selection import train_test_split
-from sklearn.metrics import classification_report, multilabel_confusion_matrix
-
-import seaborn as sns
-from pylab import rcParams
-import matplotlib.pyplot as plt
-from matplotlib import rc
 from getpaths import getpath
 
-import sys
+from bert.model import ToxicCommentTagger
+from bert.data import ToxicCommentDataModule, ToxicCommentsDataset
+from callbacks import checkpoint_callback, early_stopping_callback
 
-sys.path.append("/Users/jvkyleeclarin/Desktop/ml_finance")
 
 
+# get config
 RANDOM_SEED = 42
-
-sns.set(style="whitegrid", palette="muted", font_scale=1.2)
-HAPPY_COLORS_PALETTE = [
-    "#01BEFE",
-    "#FFDD00",
-    "#FF7D00",
-    "#FF006D",
-    "#ADFF02",
-    "#8F00FF",
-]
-sns.set_palette(sns.color_palette(HAPPY_COLORS_PALETTE))
-rcParams["figure.figsize"] = 12, 8
-
 pl.seed_everything(RANDOM_SEED)
-
-cwd = getpath("/Users/jvkyleeclarin/Desktop/ml_finance", custom=True)
-
-df = pd.read_csv(cwd / "mlfinance" / "nlp" / "toxic_comments.csv")
-df.head()
-
-
-#%%
-
-train_df, val_df = train_test_split(df, test_size=0.05)
-train_df.shape, val_df.shape
-
-# %%
-
-LABEL_COLUMNS = df.columns.tolist()[2:]
-
-# %%
-
-train_toxic = train_df[train_df[LABEL_COLUMNS].sum(axis=1) > 0]
-train_clean = train_df[train_df[LABEL_COLUMNS].sum(axis=1) == 0]
-
-
-# %%
-train_df = pd.concat([train_toxic, train_clean.sample(5)])
-
-train_df.shape, val_df.shape
-# %%
-BERT_MODEL_NAME = "bert-base-cased"
-tokenizer = BertTokenizer.from_pretrained(BERT_MODEL_NAME)
-# %%
-
-sample_row = df.iloc[3]
-sample_comment = sample_row.comment_text
-sample_labels = sample_row[LABEL_COLUMNS]
-
-print(sample_comment)
-print()
-print(sample_labels.to_dict())
-# %%
-encoding = tokenizer.encode_plus(
-    sample_comment,
-    add_special_tokens=True,
-    max_length=512,
-    return_token_type_ids=False,
-    padding="max_length",
-    return_attention_mask=True,
-    return_tensors="pt",
-)
-
-encoding.keys()
-# %%
-encoding["input_ids"].shape, encoding["attention_mask"].shape
-# %%
-encoding["input_ids"].squeeze()[:20]
-# %%
-encoding["attention_mask"].squeeze()[:20]
-# %%
-print(tokenizer.convert_ids_to_tokens(encoding["input_ids"].squeeze())[:20])
-# %%
-token_counts = []
-
-for _, row in train_df.iterrows():
-    token_count = len(
-        tokenizer.encode(row["comment_text"], max_length=512, truncation=True)
-    )
-    token_counts.append(token_count)
-# %%
-sns.histplot(token_counts)
-plt.xlim([0, 512])
-# %%
-MAX_TOKEN_COUNT = 512
-# %%
-class ToxicCommentsDataset(Dataset):
-    def __init__(
-        self, data: pd.DataFrame, tokenizer: BertTokenizer, max_token_len: int = 128
-    ):
-        self.tokenizer = tokenizer
-        self.data = data
-        self.max_token_len = max_token_len
-
-    def __len__(self):
-        return len(self.data)
-
-    def __getitem__(self, index: int):
-        data_row = self.data.iloc[index]
-
-        comment_text = data_row.comment_text
-        labels = data_row[LABEL_COLUMNS]
-
-        encoding = self.tokenizer.encode_plus(
-            comment_text,
-            add_special_tokens=True,
-            max_length=self.max_token_len,
-            return_token_type_ids=False,
-            padding="max_length",
-            truncation=True,
-            return_attention_mask=True,
-            return_tensors="pt",
-        )
-
-        return dict(
-            comment_text=comment_text,
-            input_ids=encoding["input_ids"].flatten(),
-            attention_mask=encoding["attention_mask"].flatten(),
-            labels=torch.FloatTensor(labels),
-        )
-
-
-# %%
-bert_model = BertModel.from_pretrained(BERT_MODEL_NAME, return_dict=True)
-
-#%%
-
-train_dataset = ToxicCommentsDataset(train_df, tokenizer, max_token_len=MAX_TOKEN_COUNT)
-
-sample_item = train_dataset[0]
-sample_item.keys()
-# %%
-sample_batch = next(iter(DataLoader(train_dataset, batch_size=1, num_workers=0)))
-sample_batch["input_ids"].shape, sample_batch["attention_mask"].shape
-# %%
-output = bert_model(sample_batch["input_ids"], sample_batch["attention_mask"])
-# %%
-print(output.last_hidden_state.shape, output.pooler_output.shape)
-
-#%%
-class ToxicCommentDataModule(pl.LightningDataModule):
-    def __init__(self, train_df, test_df, tokenizer, batch_size=8, max_token_len=128):
-        super().__init__()
-        self.batch_size = batch_size
-        self.train_df = train_df
-        self.test_df = test_df
-        self.tokenizer = tokenizer
-        self.max_token_len = max_token_len
-
-    def setup(self, stage=None):
-        self.train_dataset = ToxicCommentsDataset(
-            self.train_df, self.tokenizer, self.max_token_len
-        )
-
-        self.test_dataset = ToxicCommentsDataset(
-            self.test_df, self.tokenizer, self.max_token_len
-        )
-
-    def train_dataloader(self):
-        return DataLoader(
-            self.train_dataset, batch_size=self.batch_size, shuffle=True, num_workers=0
-        )
-
-    def val_dataloader(self):
-        return DataLoader(self.test_dataset, batch_size=self.batch_size, num_workers=0)
-
-    def test_dataloader(self):
-        return DataLoader(self.test_dataset, batch_size=self.batch_size, num_workers=0)
-
-
-# %%
-
 N_EPOCHS = 10
 BATCH_SIZE = 12
+MAX_TOKEN_COUNT = 512
+BERT_MODEL_NAME = "bert-base-cased"
+
+
+# get data
+cwd = getpath()
+df = pd.read_csv(cwd / "mlfinance" / "nlp" / "toxic_comments.csv")
+LABEL_COLUMNS = df.columns.tolist()[2:]
+
 
 data_module = ToxicCommentDataModule(
     train_df, val_df, tokenizer, batch_size=BATCH_SIZE, max_token_len=MAX_TOKEN_COUNT
 )
-# %%
-class ToxicCommentTagger(pl.LightningModule):
-    def __init__(self, n_classes: int, n_training_steps=None, n_warmup_steps=None):
-        super().__init__()
-        self.bert = BertModel.from_pretrained(BERT_MODEL_NAME, return_dict=True)
-        self.classifier = nn.Linear(self.bert.config.hidden_size, n_classes)
-        self.n_training_steps = n_training_steps
-        self.n_warmup_steps = n_warmup_steps
-        self.criterion = nn.BCELoss()
 
-    def forward(self, input_ids, attention_mask, labels=None):
-        output = self.bert(input_ids, attention_mask=attention_mask)
-        output = self.classifier(output.pooler_output)
-        output = torch.sigmoid(output)
-        loss = 0
-        if labels is not None:
-            loss = self.criterion(output, labels)
-        return loss, output
-
-    def training_step(self, batch, batch_idx):
-        input_ids = batch["input_ids"]
-        attention_mask = batch["attention_mask"]
-        labels = batch["labels"]
-        loss, outputs = self(input_ids, attention_mask, labels)
-        self.log("train_loss", loss, prog_bar=True, logger=True)
-        return {"loss": loss, "predictions": outputs, "labels": labels}
-
-    def validation_step(self, batch, batch_idx):
-        input_ids = batch["input_ids"]
-        attention_mask = batch["attention_mask"]
-        labels = batch["labels"]
-        loss, outputs = self(input_ids, attention_mask, labels)
-        self.log("val_loss", loss, prog_bar=True, logger=True)
-        return loss
-
-    def test_step(self, batch, batch_idx):
-        input_ids = batch["input_ids"]
-        attention_mask = batch["attention_mask"]
-        labels = batch["labels"]
-        loss, outputs = self(input_ids, attention_mask, labels)
-        self.log("test_loss", loss, prog_bar=True, logger=True)
-        return loss
-
-    def training_epoch_end(self, outputs):
-
-        labels = []
-        predictions = []
-        for output in outputs:
-            for out_labels in output["labels"].detach().cpu():
-                labels.append(out_labels)
-            for out_predictions in output["predictions"].detach().cpu():
-                predictions.append(out_predictions)
-
-        labels = torch.stack(labels).int()
-        predictions = torch.stack(predictions)
-
-        """    for i, name in enumerate(LABEL_COLUMNS):
-      class_roc_auc = auroc(predictions[:, i], labels[:, i])
-      self.logger.experiment.add_scalar(f"{name}_roc_auc/Train", class_roc_auc, self.current_epoch)"""
-
-    def configure_optimizers(self):
-
-        optimizer = AdamW(self.parameters(), lr=2e-5)
-
-        scheduler = get_linear_schedule_with_warmup(
-            optimizer,
-            num_warmup_steps=self.n_warmup_steps,
-            num_training_steps=self.n_training_steps,
-        )
-
-        return dict(
-            optimizer=optimizer, lr_scheduler=dict(scheduler=scheduler, interval="step")
-        )
-
-
-# %%
-dummy_model = nn.Linear(2, 1)
-
-optimizer = AdamW(params=dummy_model.parameters(), lr=0.001)
 
 warmup_steps = 20
 total_training_steps = 100
 
-scheduler = get_linear_schedule_with_warmup(
-    optimizer, num_warmup_steps=warmup_steps, num_training_steps=total_training_steps
-)
-
-learning_rate_history = []
-
-for step in range(total_training_steps):
-    optimizer.step()
-    scheduler.step()
-    learning_rate_history.append(optimizer.param_groups[0]["lr"])
-# %%
 steps_per_epoch = len(train_df) // BATCH_SIZE
 total_training_steps = steps_per_epoch * N_EPOCHS
-# %%
+
 warmup_steps = total_training_steps // 5
 warmup_steps, total_training_steps
-# %%
+
+
 model = ToxicCommentTagger(
     n_classes=len(LABEL_COLUMNS),
     n_warmup_steps=warmup_steps,
     n_training_steps=total_training_steps,
 )
-# %%
-criterion = nn.BCELoss()
-# %%
-checkpoint_callback = ModelCheckpoint(
-    dirpath="checkpoints",
-    filename="best-checkpoint",
-    save_top_k=1,
-    verbose=True,
-    monitor="val_loss",
-    mode="min",
-)
-# %%
-logger = TensorBoardLogger("lightning_logs", name="toxic-comments")
-# %%
-early_stopping_callback = EarlyStopping(monitor="val_loss", patience=2)
-# %%
+
 trainer = pl.Trainer(
-    logger=logger,
     checkpoint_callback=checkpoint_callback,
     callbacks=[early_stopping_callback],
     max_epochs=N_EPOCHS,
     gpus=0,
     progress_bar_refresh_rate=1,
 )
-# %%
-trainer.fit(model, data_module)
-# %%
 
+
+trainer.fit(model, data_module)
 
 trainer.save_checkpoint(filepath)
