+++ mlfinance/nlp/train.py
@@ -3,11 +3,15 @@ from omegaconf import DictConfig
 
 
 
+def main(cfg):
+    print(cfg)
+
+
 
 @hydra.main(config_path="./conf", config_name="config")
 def hydra_entry(cfg: DictConfig) -> None:
-    print(cfg)
-    pass
+    main(cfg)
+
 
 
 if __name__ == "__main__":
